feat: add toast notification & fix error messages

+++ frontend/src/components/WorkoutDetails.jsx
@@ -3,6 +3,8 @@ import { useWorkoutsContext } from "../../hooks/useWorkoutsContext"
 import { formatDistanceToNow } from 'date-fns'
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
+import { Bounce, toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 
 const WorkoutDetails = ({ workout }) => {
 
@@ -10,6 +12,7 @@ const WorkoutDetails = ({ workout }) => {
     const navigate = useNavigate();
     const { dispatch } = useWorkoutsContext()
     const [isEditing, setIsEditing] = useState(false)
+    const [isDeleting, setIsDeleting] = useState(false)
     const [title, setTitle] = useState(workout.title)
     const [load, setLoad] = useState(workout.load);
     const [reps, setReps] = useState(workout.reps);
@@ -18,25 +21,44 @@ const WorkoutDetails = ({ workout }) => {
     const handleTitleClick = () => {
         navigate("/test", { state: { workout } });
     }
-
     const confirmDelete = () => {
         setDeleteConfirm(true);
     }
+    const handleCancelEdit = () => {
+        setIsEditing(false);
+        setTitle(workout.title);
+        setLoad(workout.load);
+        setReps(workout.reps);
+        toast.info("Editing canceled", { autoClose: 2000, hideProgressBar: true, transition: Bounce });
+    };
+    const handleEnableEdit = () => {
+        setIsEditing(true);
+        toast.info("Editing Enabled", { autoClose: 2000, hideProgressBar: true, transition: Bounce });
+    };
 
     const handleDelete = async () => {
+        if (isDeleting) return
+        setIsDeleting(true);
         try {
             console.log("Deleting workout with ID:", workout._id);
             const response = await axios.delete(`http://localhost:4000/api/workouts/${workout._id}`)
             dispatch({ type: 'DELETE_WORKOUT', payload: workout._id })
-            console.log(response.data.message);
+            toast.success(response.data.message);
         } catch (error) {
-            throw new Error(error?.response?.data?.message || "Cannot delete this workout")
+            const errorMessage = error.response?.data?.message || "Failed to delete workout!";
+            toast.error(errorMessage);
+            throw new Error(errorMessage)
         }
         setDeleteConfirm(false);
     }
 
     const handleUpdate = async () => {
+        if (title.trim() === '') {
+            toast.error("Title cannot be empty!");
+            return setError("Title cannot be empty!");
+        }
         if (load < 0 || reps < 0) {
+            toast.error("Load and reps cannot be negative!");
             return setError("Load and reps cannot be negative!");
         }
 
@@ -49,9 +71,11 @@ const WorkoutDetails = ({ workout }) => {
             dispatch({ type: 'UPDATE_WORKOUT', payload: response.data.updateWorkout })
             setIsEditing(false)
             setError(null)
-            console.log(response.data);
+            toast.success(response?.data?.message);
         } catch (error) {
-            throw new Error(error?.response?.data?.message || error.message)
+            const errorMessage = error.response?.data?.message || "Failed to update workout!";
+            toast.error(errorMessage);
+            throw new Error(errorMessage)
         }
     }
 
@@ -67,30 +91,30 @@ const WorkoutDetails = ({ workout }) => {
 
 
                         <input className="border p-1 rounded-lg border-gray-300 mr-2" placeholder="Workout Title" value={title} onChange={(e) => setTitle(e.target.value)} />
-                        <input className="border p-1 rounded-lg border-gray-300 mr-2" type="number" min='1' placeholder="Load"  value={load} onChange={(e) => setLoad(e.target.value)} />
+                        <input className="border p-1 rounded-lg border-gray-300 mr-2" type="number" min='1' placeholder="Load" value={load} onChange={(e) => setLoad(e.target.value)} />
                         <input className="border p-1 rounded-lg border-gray-300 mr-2" type="number" min='1' placeholder="Reps" value={reps} onChange={(e) => setReps(e.target.value)} />
                         <p className="">Last updated: {formatDistanceToNow(new Date(workout.updatedAt), { addSuffix: true })} ago</p>
                         <div className="flex mr-3">
-                            <svg onClick={handleUpdate} xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="p-1 rounded-lg bg-green-600 text-white hover:bg-green-500 active:bg-green-600 mr-2 size-10">
+                            <svg onClick={handleUpdate} xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="p-1 rounded-lg bg-green-600 text-white hover:bg-green-500 active:bg-green-600 mr-2 size-10 cursor-pointer">
                                 <path strokeLinecap="round" strokeLinejoin="round" d="m4.5 12.75 6 6 9-13.5" />
                             </svg>
                             {/* cancel button */}
-                            <svg onClick={() => { setIsEditing(false) }} xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="p-1 rounded-full bg-gray-300 text-white hover:bg-gray-400 active:bg-gray-300 mr-2 size-10">
+                            <svg onClick={handleCancelEdit} xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="p-1 rounded-full bg-gray-300 text-white hover:bg-gray-400 active:bg-gray-300 mr-2 size-10 cursor-pointer">
                                 <path strokeLinecap="round" strokeLinejoin="round" d="M6 18 18 6M6 6l12 12" />
                             </svg>
                         </div>
                     </div>
                 </div>
             ) : (<>
-                <h4 className="text-2xl font-extrabold text-blue-900 cursor-pointer hover:text-blue-700" onClick={handleTitleClick}>{workout.title}</h4>
+                <h4 className="text-2xl font-extrabold text-blue-600 cursor-pointer hover:text-blue-700" onClick={handleTitleClick}>{workout.title}</h4>
                 <div className="flex justify-between w-full pr-5 items-center">
                     <p><strong>Load (kg) : </strong>{workout.load}</p>
                     <p><strong>Reps : </strong>{workout.reps}</p>
                     <p><strong>Created : </strong>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
-                    
+
 
                     <div className="flex gap-2 mt-3">
-                        <svg onClick={() => { setIsEditing(true) }} xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="bg-blue-600 text-white rounded-lg p-2 cursor-pointer hover:bg-blue-500 transition size-10">
+                        <svg onClick={handleEnableEdit} xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="bg-blue-600 text-white rounded-lg p-2 cursor-pointer hover:bg-blue-500 transition size-10">
                             <path strokeLinecap="round" strokeLinejoin="round" d="m16.862 4.487 1.687-1.688a1.875 1.875 0 1 1 2.652 2.652L6.832 19.82a4.5 4.5 0 0 1-1.897 1.13l-2.685.8.8-2.685a4.5 4.5 0 0 1 1.13-1.897L16.863 4.487Zm0 0L19.5 7.125" />
                         </svg>
 
@@ -110,13 +134,13 @@ const WorkoutDetails = ({ workout }) => {
                         <div className="flex justify-end space-x-4 mt-4">
                             <button
                                 onClick={() => setDeleteConfirm(false)}
-                                className="px-4 py-2 bg-gray-300 hover:bg-gray-400 active:bg-gray-300 rounded-lg"
+                                className="px-4 py-2 bg-gray-300 hover:bg-gray-400 active:bg-gray-300 rounded-lg cursor-pointer"
                             >
                                 Cancel
                             </button>
                             <button
                                 onClick={handleDelete}
-                                className="px-4 py-2 bg-red-500 text-white rounded-lg hover:bg-red-600"
+                                className="cursor-pointer px-4 py-2 bg-red-500 text-white rounded-lg hover:bg-red-600"
                             >
                                 Confirm
                             </button>
